Use openapi.yaml again
Migrate from LambdaRestApi to SpecRestApi.
For each endpoint in openapi.yml, we add the lambda integration using
OpenAPI extensions for API Gateway.
Lambda authorier and request validator are also defined in openapi.yaml.

The openapi.yml file in our codebase will be the extended version of the
one on stoplight (because we shouldn't expose the aws things to public).

Whenever we update the schema in stoplight, we need to manually merge
the changes to the codebase. It can be done by
1. Download the file from https://stoplight.io/api/v1/projects/flagright/flagright-api/nodes/reference/Flagright-API.yaml
2. vimdiff Flagright-API.yaml openapi.yaml
3. Apply the new changes