feat: replace login with Auth0 universal login
If a user is not logged in, accessing any console path will
be redirected to Auth0 login page.
The current login component is the default design. We can
customize it in the future.
The real logged-in user's name will be displayed on the top bar.
Also demonstrate adding JWT (from Auth0) in the request headers
when requesting console APIs (this step will be transparent
after we have proper frontend SDK)