Save rules result along with the transaction primary item
The rules result will be saved as the two additional attributes
- `executedRules` and `failedRules` of a primary transaction
item. There's a hard limit of 400KB for a DDB item, so we cannot
save unlimited rules result. But it can go a long way (can at
least support a couple hundreds of rules).

Also try/catch `rule.computeRule` and return `failedRules`.