fix: enable importing a big file
Currently, if the import takes more than 29 seconds, the api will fail,
but the lambda doing the import will keep running.
If `api.postImport` fails, we ignore the error and poll for the import
status every 10 seconds.

NOTE: There's still a 15 min limit for the lambda running time, so we
cannot import a file that takes more than 15 minutes to import.