Add file import status
There's a hard timeout 25 secs limit of API Gateway. We need an async
way to check for the import status.
Add a new endpoint for getting the status of a file import.
We'll create a new file import record before importing and update the
import status after the import is done (success or failed).