Update rule: transaction velocity
For Icrypex rule R-29
Add two optional parameters - userIdsToCheck, checkTimeWindow.
Introduce `filters` concept in the rule implementation to allow
a rule only to be run for certain conditions.
In the future, we could potentially have a "filters library" and
allow users to apply arbitrary number of filters to any rule.