User event and user rules
Add a new endpoint for posting a user event. Rules with `USER` type
will be run against the new user event. Similar to transaction, the
rules result (`executedRules`, `failedRules`) will be saved along with
the user event in DynamoDB. Similar to transaction, a user event will
be streamed to MongoDB with an additional field `status` which is the
aggregated status of the executed rules.

New public endpoints - `POST /events/user`
New models:
- `UserEvent`
- `UserEventWithRulesResult`
- `UserMonitoringResult`