Implement rule - transactions volume (quantiles)
Same logic as `transactions-volume` rule, but instead of using arbitrary
time window, we define three different fixed quantiles - DAILY, MONTHLY, YEARLY
The transactions will only be summed together if the timestamps correspond to
the same time label.
For example, if the monthly limit is 100 USD, then 99 USD on 2022-01-31 and
99 USD on 2022-02-01 won't break the limit because they are in different months.

Instead of retreving all the transactions in the past day/month/year, we add
another user aggregation data -

Key -
PartitionKeyID: tenantId1#aggregation#user:userId1#time,
SortKeyID: 2022-01,

Attributes -
- sendingTransactionsVolume: { transactionAmount: 100, transactionCurrency: USD }
- receivingTransactionsVolume: { transactionAmount: 100, transactionCurrency: USD }

The currency of the aggregated amount is determined by the first transaction currency
for the target quantile.