Get transactions/user-events by a time range
Previously we get the transactions/user-events after a timestamp based
on the assumption that the transactions/user-events are coming in
in-order. But whenever they're imported out-of-order, the rules
using the repository methods will have incorrect result as they'll get
the transaction/events after the transaction/event's timestamp.
Change to get the transactions/user-events by time range.