Implement rule R-88 (#148)
* Allow setting initial transaction state other than CREATED

* Aggregate for a specific transaction state

For some aggregated data, we only want to update it when a transaction
is in a certain state. This allows the aggregation step to be run
multiple times for different transaction states.
For now, we make all the current aggregators only aggregate when
a transaction is in SUCCESSFUL state. If there're multiple transaction
events with the same transaction state, only one aggregation will be
run.

* Use dynamic import to import fast-geoip for typings

* Add more granularity to user's aggregated transaction countries

Before we only save the destination country for the sender and origin
country for the receiver.
Now we also save the origin country for the sender and destination
for the receiver.

* Implement rule - ip-address-unexpected-location

If the country of the sender ip address is not in
1. User's country of residence
2. User's country of nationality
3. The countries of the user's past approved transaction (origin country)

then we flag the transaction.