Fix transactions rules for transactions with anonymous origin/destination user (#271)
* Allow persisting user transaction index if sender/receiver user cannot be identified

Previously, we only persist a user transaction index if both sender and receiver can be
uniquely identified (using user ID or payment details). Now we have a case that
we still want to flag a transaction when origin user info is totally empty (without
user ID and payment details). In this case, we persist the user transaction index
without senderKeyId/receiverKeyId being set.

* Fix falsely grouped anonymous users

If origin/destination user ID is empty, we use paymentDetails to create a
unique identifier for the anonymous user. If, say, for the CARD payment method,
cardFingerprint is undefined, instead of generating the key - `...cardFingerprint:undefined...`
and persist it, we just return undefined for the whole key.
Then we don't falsely get the transactions of one anonymous user from the transactions of another
anonymous user.

* Remove destinationUserId requirement for transaction-velocity rule

* Make transaction-volume rule support anonymous user

Follow the same as transaction-velocity.