Make tarpon-change-capture-kinesis-consumer work in local env (#315)
For the entities that should be handled by the kinesis consumer, we explicitly call the consumer
after we update those entities. Add `localTarponChangeCaptureHandler` that will create a
kinesis event object and call tarpon-change-capture-kinesis-consumer. In this way, we don't need
to set up kinesis stream locally. We investigated using localstack to spin up local DynamoDB and
Kinesis stream, but there're some weird connection issues. This is the stopgap solution to unblock
developing the critical flow locally.

Context:
Previously, we're not running DynamoDB Stream and Kinesis Stream locally, so
tarpon-change-capture-kinesis-consumer cannot be triggered, thus we cannot test it locally.
The only was to test it was deploying to Dev AWS account and test there. This dragged down the
development speed a lot.

Local flow:
Lambda -> DynamoDB

Cloud flow:
Lambda -> DynamoDB -> DynamoDB Stream -> Kinesis Stream -> tarpon-change-capture-kinesis-consumer

End-to-end local dev flow:
After this change, we can create a new transaction/user from API and see it in console.
We need to explicitly set the tenant ID by passing the `Tenant-Id` http header.
e.g
```
curl --request POST \
  --url http://localhost:3000/transactions \
  --header 'Content-Type: application/json' \
  --header 'Tenant-Id: flagright' \
  --data '{
  "transactionId": "123",
  "originUserId": "8650a2611d0771cba03310f74bf6",
  "destinationUserId": "9350a2611e0771cba03310f74bf6",
  "originAmountDetails": {
    "country": "DE",
    "transactionAmount": 800,
    "transactionCurrency": "EUR"
  },
  "destinationAmountDetails": {
    "country": "IN",
    "transactionAmount": 68351.34,
    "transactionCurrency": "INR"
  },
  "promotionCodeUsed": true,
  "timestamp": 1641654664000,
  "originPaymentDetails": {
    "method": "CARD",
    "cardFingerprint": "20ac00fed8ef913aefb17cfae1097cce",
    "cardIssuedCountry": "US",
    "transactionReferenceField": "Deposit",
    "3dsDone": true
  },
  "destinationPaymentDetails": {
    "method": "CARD",
    "cardFingerprint": "20ac00fed8ef913aefb17cfae1097cce",
    "cardIssuedCountry": "IN",
    "transactionReferenceField": "Deposit",
    "3dsDone": true
  }
}'
```