Switch DynamoDB to on-demand mode in production (#325)
* Switch DynamoDB to on-demand mode in production

We noticed that Nextpay experienced `ProvisionedThroughputExceededException`
when they send many transactions in a short period of time.
Some transactions were eventually processed successfully after DynamoDB SDK's
built-in retry mechanism (but the http request still failed because of
API Gateway's 30 sec timeout). And some transactions were eventually failed
to be processed because it reached the max retries and those transaction were
lost.
If we stay with the provisioned mode, we'll need to fine-tune the provisioned
capacity and turn on auto-scaling to ensure the incoming requests have equal
DynamoDB access performance.
To make it simplier for now, we turn on on-demand mode for production (higher
cost) and create alams (ConsumedReadCapacityUnits, ConsumedWriteCapacityUnits)
to monitor the usage. After we scale and learn more about our request pattern,
we could switch back to the provisioned mode for cost saving.

* Use Maximum statistics