Fix Table failing to reload (#294)
Fix regression of 2570886.
Some Table requests depend on the caller state. We used to create a new callback
on every re-render and it'll use the latest state. But we refactored to put the
request prop to a Table's internal initial state. If the request's dependant
data changes, it won't be used by Table.
Change to pass request as a memoized callback. Whenver the dependancies of request
change, a new request will be triggered.
For manually reloading the table, we implement our own reload method in Table, because
the original `actionRef.current.reload()` doesn't trigger a new request probably due to
we made Table a controlled component (passing `dataSource` by our own).

This fixes:
1. Changing date range in dashboard table widgets doesn't work
2. Allowing a case doesn't reload the table