Fix beforeTimestamp DynamoDB query (#358)
According to the documentaion:
```
a BETWEEN b AND c — true if a is greater than or equal to b, and less than or equal to c
```
But we don't want to get the transaction with timestamp == beforeTimestamp because the
current transaction should be excluded. We could incorrectly fetch one more transaction
when verifying a transaction event and flag the transaction incorrectly.