Migrate transactions referencing to the deleted rules (#382)
Iterate over all transactions of a tenant. For each transaction, if
its executedRules/hitRules includes the deleted rules, we change the
ruleId to the redirected one.

Considering our total transaction volume is still not too big and we
shouldn't do this kind of migration anymore (a rule shouldn't be deleted).
This is accetable for now.