Send password change email when a new user is invited (#407)
* Send password change email when a new user is invited

After a user is invited, the user will receive two emails - one for verifying
the email and one for setting the new password. The initial password
is a random uuid. This way the inviter doesn't know the password of the
invited user so s/he cannot log in as the invited user.

* Rollback user creation if failing to add user to org