Kinesis consumer retry mechanism (#417)
* Kinesis consumer retry mechanism

If error occurs in a kinesis consumer, for each dynamodb update, we'll send
a new kinesis event to the "retry stream" belonging to the consumer. For each
dynamodb update, we know it's entity ID (e.g transaction ID). And we'll add
the entity key to another DynamoDB table 'Transient', then for the future updates
related to the entity ID, we'll directly send them to the retry stream to keep
the events order (e.g if transaction event E1 for transaction T1 fails and is sent
to the retry stream, event if event E2 for transaction T1 won't fail, but we'll still
send it to the retry stream. Otherwise, E2 might complete earlier than E1).
This retry mechanism isolates the failed entities out and prevent them from blocking
the futher events for other entities. After a event is sent to the retry stream, it'll
probably also fail, but we'll have alarm for it, and we'll fix it asap and deploy the fix,
and the events in the retry stream should be successfully processed eventually.

New components:
1. One additional kinesis retry stream for each consumer
2. DynamoDB "Transient" table.

Old Architecture:
```
DynamoDB Stream -> Kinesis Stream -> Consumer
(if fails to process one event, will keep retrying, and newer events will be blocked)
```

New Architecture:
```
DynamoDB Stream -> Kinesis Stream         -> Consumer --------> DynamoDB "Transient" Table
		   Kinesis Stream (retry) ->	      | (Send to retry stream if necessary)
				     ^----------------
```

Note that the kinesis consumer now has two event sources (one from the original stream and one
from the retry stream). We determine is the event if from the retry stream or not by checking
`eventSourceARN` to see if the retry stream name is in it.