Fix openapi deployment (#448)
In 944db56, we don't add lambda integration to `parameters` because it's not a
HTTP method (It's at the method level because all the methods share the same
path parameters). But it causes failing to deploy to API Gateway because of
missing lambda integration.
According to https://docs.aws.amazon.com/apigateway/latest/developerguide/http-api-open-api.html,
path parameters should live under method setting. In the openapi augmentor,
if we see a path-level parameters, we concat it to the parameters in each method
and delete it from the path-level.