Retry DynamoDB operation with refreshed credentials for migration (#566)
When running the migrations, we'll assume CodePipelineDeployRole role
in a different account (dev,sandbox,prod) to get the temporary credentials
for accessing DynamoDB in each account. Because of role chaining, we can
only have max 1 hour session duration. And it seems we cannot use cross-account
role as the role for CodeBuild.
Instead, we live with 1 hour max session duration and we re-assume the cross-account
role every 50 minutes (before the credentials expire). And if a DynamoDB operation
fails, we retry with the refreshed credentials.

Revert the changes in #555 for setting duration-second.