Enable uploading documents when closing a case (#583)
When a user closes cases with uploaded documents, we create a case comment
with the documents as the attached files for each case. As the files will
be copied from the tmp bucket to the permanent bucket for each comment, the
files in the comments are independant from each other depite that they're
the same files.