Sentry source maps support (#629)
For each lambda bundle code (app.js), we generate a source map file for it
(app.js.map) when running `npm run build`. In the "Build" code pipeline stage,
we clear all the files for `latest-version` release version first and re-upload
the files (both app.js and app.js.map). `latest-version` release version is used
by Dev/Sandbox only as we don't really care about releases of these envs. And we
could have too many releases because we'll create one for each new commit.
For production, we use the latest commit hash as the release version.

The files uploaded to sentry is in the format `~/${code_path}/app.js` and
`~/${code_path}/app.js.map`. In order to let Sentry find the corresponding
source maps, we add a prefix `app:///${code_path/` using RewriteFrames.

As we use webpack to generate source maps, it'll make codebuild OOM with 15GB memory.
We change to run webpack in 2 batches - 1 batch for half lambdas and 1 batch
for the other half.