Fix user case details page crash for kevin (#638)
* Fix pattern velocity/percentage rules

Only run the rule when the incoming transaction matches the pattern.

* Fix user case pipeline OOM when there're too many transactions

Then 16MB limit also applies to aggregation pipeline. When there're
more than 1000 transactions in a case, we slice it to only have 1000
transactions to lookup the actual transactions.

We'll limit the max transactions we can put to a case in FDT-85722.