Optimise cases page loading time (#671)
* Restore materialized caseTransactions in the cases collection

We used to store the actual transaction payload in a case (in caseTransactions), but we
removed it when fixing a user case issue #558 that we were storing too many transactions
in a single case and reached the max 16MB document limit. For TRANSACTION case, we need
to save the actual caseTransactions because we need to display transaction info in the
case table. Relying on $lookup will literally scan all the cases, and that makes loading
cases page way too slow when there're many cases in total.

Changes
- Save `caseTransactions` to ${tenantId}-cases mongo collection (migration added)
- Remove the aggregation stage for transactions lookup
- Split the aggregation pipeline for fetching cases to 2 parts - 1 for the stages before
  applying $limit and 1 for the stages after applying $limit.
- Run includeTransactionUsers stage after $limit (except for sorting user name in
  transaction cases)
- Run includeTransactionEvents stage after $limit as we don't need to filter/sort on events
- Add one more index for (caseStatus, caseType, createdTimestamp)

* Create a new user case if the existing one has 1000 transactions

If the existing user case already has 1000 transactions, we don't add
the new transaction to it. Instead, we create a new case.

* Revert "Remove CDK SSO Sync (#643)"

This reverts commit 36c51c5091adfc032c5e52ef10f97bd5b239c164.