Use SQS queues instead of Kinesis stream for retry (#782)
* clean up

* Use SQS queues instead of Kinesis stream for retry

In edaf6c7, we create two streams (one main stream and one retry stream) for a
kinesis stream consumer. The retry stream is mostly idle and empty but it still
has costs. As failed records don't need to (and won't) be processed in real-time as
we probably need to deploy a fix to consume thme successfully, moving to failed records
to a SQS queue is more cost-effective and pragmatic.

Delete retry kinesis streams and replace with FIFO SQS queues. A failed kinesis record
will be sent to the SQS retry queue. Records with the same entity ID will be processed
in order, while records with different entity IDs could be processed concurrently and
out-of-order.

Old Architecture:
```
DynamoDB Stream -> Kinesis Stream         -> Consumer
                   Kinesis Stream (retry) /         | (Send to retry stream if necessary)
                                     ^--------------
```

New Architecture:
```
DynamoDB Stream -> Kinesis Stream         -> Consumer (kinesis)-
                   Retry SQS Queue        -> Consumer (sqs)     | (Send to retry queue if necessary)
                                ^--------------------------------
```