Ensure API Gateway response template won't be reset (#785)
All our previous attempts to fix the missing validation response template
(ed5db61, 53268c5, 53cc84a) are to modify the content of the template.
While the root cause of the template being reset after deployment is unknown,
as a workaround, we do the hack fix as we did before automatically. We append
random number of spaces (1-1000) at the end of the template string to make
cdk deployment always deploy the template on every deployment. The spaces suffix
won't affect the actual response payload.