Optimise MongoDB aggregation pipeline for getting unique transaction fields (#799)
* Change all MongoDB regex match to be prefix case-sensitive match

Arbitrary regex match couldn't use the indexes efficiently. If we don't add
"prefix search" and "case-sensitive" constraints, We could end up with full
documents/index keys scan which is bad.
Although now users cannot search a substring which isn't prefix string, it's
a necessary sacrifice to make filtering performance in Console.

* Optimize get uniques for transaction fields

If we don't have the filter conditions, we start with $group instead of
$match, otherwise we'll just scan all the documents.
If we do have the filter conditions, we start with $match but we add an
additional { $limit: 10000} to cap the max documents to filter. It's
acceptable because it's for auto-complete and we don't need 100% correctiness.