Allow local to fetch secrets from Dev account (#802)
Change all secret manager usage to use the v3 SecretsManagerClient SDK.
If the dev credentials in `~/.aws/credentials` is valid, local tarpon will
fetch the secrets in the Dev account. If the credential is expired, an error
will be thrown and there'll an error message to ask for runnin `npm run aws-sso-login:dev`
to refresh the credentials.