Flatten openapi allOf to improve validation error message (#813)
AWS APIGateway request validator detail error messagae doesn't work well for `allOf`.
For example, `allOf[A, B]`, and we have a missing required parameter in A,
the error message will just say:
"instance failed to match all required schemas (matched only 1 out of 2)"
This is the workaround to "flatten" the schema (recursively) to have better error message.

Note that the same problem still happen if we have `oneOf`. But for public API for now,
only PaymentDetails uses `oneOf`. If the error happens not in PaymentDetails, we'll get
the proper actionable error message.