Fix api key authorizer and add missing indexes (#853)
* Add missing mongodb indexes

* Return placeholder response in api-key authorizer if api key is invalid

If the api key is invalid, lambda authorizer will still be executed and the key
will be validated later by APIGateway. To avoid false alert, if we cannot
parse the api key, we return the placeholder response. Even we can parse
the invalid api key, it'll be rejected later by APIGateway.

* Update postman test