Apply rule aggregation to new rules and update rule aggregation logic (#885)
* Apply aggregation to rule high-traffic-between-same-parties

* Apply aggregation to rule - high-traffic-volume-between-same-users

* Apply aggregation to rule - high-unsuccessfull-state-rate

Also fix the issue that the aggregation user key ID shouldn't
have `direction` which will cause incorrect rule result.

* Add hard-coded rule aggregation version per rule

The hard-coded rule aggregation version is used along with the dynamic version
(from getAggregationVersion()). We need to bump the version whenver we update the
rule aggregation implementation if it'll make the existing aggregated data invalid.

* fix test

* Update aggregation - transactions-average-exceeded

Implement `computeRuleUser` to handle origin/destination user separately,
then it's possible to use aggregation data for origin user and use raw
data for destination user. Previously, if one of origin/destination users
doesn't have aggregation data, we get raw data for both users. This impacts
the performance when one of the user is a new user with historical transactions.

* Update aggregation - transactions-exceed-past-period

* Update aggregation - transactions-pattern-percentage-base

---------

Co-authored-by: Madhu G Nadig <madhug.nadig@gmail.com>