Support multiple allowed origins for CORS (#891)
The 'Access-Control-Allow-Origin' response header can only be one value.
Previously we hard-code the values as the response header parameter and we
can only support a single url per env. To support white-label customers (with
its own console url), we use the apache velocity template to check if the
origin is in the allowed list of origins, if true, we return the
'Access-Control-Allow-Origin' header as the input origin.