Increase kinesis event source batch size to the max value (#897)
Change from batch size 10 -> 10000 (max).
For dynamodb update we need to process, if it fails, we'll send it
to the retry sqs queue without throwing an error. Thus we can have
a bigger batch size without worring about failure that causes the whole
batch to be retried.
All set max batching window to 10 seconds, which implies that it'll wait
for either we have 10000 records or 10 seconds to trigger the lambda.