Fix Kevin incorrect rule R-125 hits (MongoDB only) (#918)
Fix Kevin incorrect rule hits in MongoDB

Related incident: https://github.com/flagright/tarpon/pull/906
This is the "shallow-fix" to make R-125 not hit for the affected time interval.
The transactions with R-125 hit:
    - 02/01 -> 02/13: 29
    - 02/13 -> 02/17: 52871
The actual R-125 hit during the affected time interval should just be a few. We
sacrifice those false negatives to eliminate the many false positives.
We'll be able to re-run rules for past transactions properly in FDT-47790 -
https://www.notion.so/flagright/Rerun-rules-for-existing-transactions-for-backfixing-facf9dad03794b9aa675ddda07cecc1d?pvs=4

NOTE: Affected transactions in DynamoDB are not updated as the current tarpon kinesis consumer
is not ready to handle this case and could create new cases unexpectedly. As the chance of those transactions being
updated in DynamoDB is low, we keep them untouched to make the fix simpler.

Co-authored-by: Madhu G Nadig <madhug.nadig@gmail.com>