Add new optional parameters to R-119, R-126 (#1034)
* Minor refactor high-traffic rules

To follow the current convention of implementation:
```
computeRule() {
  const data = await this.getData()
  ...
}

getData() {
   // get data from aggregation
   // if aggregation data not found, get from raw transactions
   // rebuild aggregation from the raw transactions
   // return data
}
```

* Add optional parameters to high-traffic rules

If `originMatchPaymentMethodDetails` is true, the sender is identified by
the origin payment details.
If `destinationMatchPaymentMethodDetails` is true, the receiver is identified
by the destination payment details.

* Update partition key for GENERIC_BANK_ACCOUNT

Previously we only use `accountNumber` and `accountType` to build the
identifier. But Kevin only have `accountNumber` for destination user
and `bankId` and `paymentChannel` for origin user. We change to concat
all of accountNumber, accountType, bankCode, paymentChannel as the
identifier.

* Change integer to number

* Add a script for running local rules engine using remote transactions

* fix test

---------

Co-authored-by: Madhu G Nadig <madhug.nadig@gmail.com>