Handle cases creation for user rules hit (#1193)
We do the same thing as transactions for users. When we run user rules for a user,
`exectuedRules`/`hitRules` will be persisted to the user object in DynamoDB. And in
the tarpon change kinesis consumer, we create new cases (or merge to the existing ones)
by using `user.hitRules`.

Also add `ruleHitMeta` to `Alert` model. It'll be used to display sanctions info for an Alert.