Destroy dynamodb underlying resources at the end - 2 (#1237)
Fixes e43c57cac. We forgot to add `await` before the handler, so the code inside
`finally` gets run prematurely (clean up nothing).

Also, instead of putting `cacheDynamoDbClients` in the global context, we put it
inside the lambda call context to make sure it won't be reused across the calls.