Fix missing cases for R-169 hit  (#1372)
* Don't set custom _id for a case

We're not querying by `_id`. Let it be automatically generated.

* Fix missing cases for R-169 hit

The hit direction of R-169 should be reversed. We only check the counterparty without user ID.
If the destination user is hit, we should flag origin user, and vice versa.
Currently we only flag the unknownn counterparty and as we only create cases for a known user,
so no cases are created for R-169.