Avoid rebuilding rule aggregation unnecessarily (#1820)
If we get empty data from the rule aggregation data for the target time window,
we check if there's any aggregation data for the partition key. If true, it implies
that the aggregation data was rebuilt before and we can safely return [] (instead of
undefined) to avoid rebuilding the aggregation data completely.
The above case happens when the target time window is without rolling basis - When we process
the first transaction after 00:00 and try to get the aggregation data, we'll get nothing.