Improve qa-stack-cleanup script (#1981)
We cannot delete too many stacks at the same time as the deletion could fail due
to the rate limit.
Change to delete them one by one (we only delete the next one when the previous one is
deleted successfully). And we'll retry up to 3 times if it somehow fails to delete.