Update rule aggregation synchronously if the latest version is rebuilt (#2131)
* Update rule aggregation synchronously if the latest version is rebuilt

Before this change, we'll always rebuild and update aggregation in the transaction lambda, and
we'll always rebuild and update aggregation in the transaction aggregation lambda if RULES_ENGINE_V2
is enabled.
For now, as we fetch all the required transactions into memory when rebuilding the aggregation, we
need to set either the transaction lambda or aggregation lambda to have big memory and it'll be triggered
for every single transaction. It's not cost effective as we only rebuild when there's a version change, and
most of the time we just do simple low-memory aggregation update, thus wasting the memory resource and money.

This change makes aggregation update run synchronously inside the transaction lambda is the latest version
is already rebuilt, and we only send the aggregation tasks to the aggregation lambda when the latest version
is not ready. In this way, we can keep the transaction lambda to have low memory and the aggregation lambda with
big memory will only be triggered when we need to rebuild the aggregation (and updating aggregations if the
rebuild is in-progress).

Advantages:
- Lower cost
- More up-to-date aggregation

Disadvantages:
- Just a bit higher latency (updating aggregation should be relatively fast)

* Change transaction lambda memory back to 1G for prod:eu-1