Workarounds of public API timeout (#2163)
* Increase hard timeout to 15 min for the lambdas that run rules

It looks like when we're rebuilding aggregations which queries Tarpon tables (for
raw transactions) a lot, the transaction lambda's dynamodb query will be affected
(long query time probably because of SDK's retries).
This is a quick fix to prevent the lambda from being killed after 100 seconds and
make it complete the work.

* Rebuild rule aggregation for a single tenant sequentially

Previously we rebuild the rule aggregation for different users in parallel and
it seems to cause too much DynamoDB read and downgrades the query performance
a lot. This is a quick workaround to make sure we only rebuild for a single user
at a time and reduce the DynamoDB read.