Refactor DashboardRefreshBatchJobRunner (#2295)
* Add updatedAt to InternalTransaction

* Refactor DashboardRefreshBatchJobRunner to cover entity updates

For case and transaction stats, as we're using the created timestamp
for grouping, we need to get the affected hours and refresh them individually.

For example:

               |''''''''''| <- Check time range
---------------------------------------> Time
^                  ^
C-1 created        C-1 updated

Previously, C-1 update won't be relected in the dashboard because we filter
the cases by createdTimestamp.

Now, we get the cases updated during the check time range first and construct
the affected time ranges. It could be 1-N ranges.

The same applies to transaction stats.

For team stats, as we're using the status change timestamp, we only need to use the
check time range.

Also remove all the places that trigger the dashboard refresh job except for the 10-min
cron job handler.

* Migration: add updatedAt to internal transactions

* Migration: refresh all dashboard stats

* Clean up stale dashboard data after refresh

After we refresh dashboard data for a target date range, we add
`updatedAt` to the refreshed records. For the records within the
target date range with `updatedAt` lower than now, we remove them.

* Fix "Top users by rule hits" dashboard widget "Open cases"