Only send transaction aggregation tasks when update is needed (#2336)
Add a new method `shouldUpdateUserAggregation` to `TransactionAggregationRule`.
If it returns true, it implies we need to use the transaction to update a user's
rule aggregation data.

Context (banked-uk):
For a transaction and N active rule instances, we could send up to 2N aggregation tasks
(origin and destination). And if some rule instances have transaction historical filters
set, for those transactions not filtered, we're stilling sending aggregation tasks and the
task will do nothing, and we'll keep doing it because `TransactionAggregationRule::isRebuilt`
will always return false because we don't rebuild for those transactions.
And for each aggregation task, we'll read the transaction/user/rule from DynamoDB again
and it leads to unnecessary DynamoDB reads.