Fix transactions 1ms before not visible to the current transaction (#2379)
Regression of fd4a5898a00. Previously we find the transactions before the current
transaction by setting `beforeTimestamp` to be the timestamp of the current transaction
minus one because we don't want to match the current transaction (if run multiple times).
But after fd4a5898a00 we append transaction ID to the sort key of the auxiliary transacions,
making the minus one unnecessary as `{this.transaction.timestamp}` is always smaller than
`{this.transaction.timestamp}-{this.transaction.transactionId}`. So even if DynamoDB BETWEEN
to is inclusive, we won't match the current transaction.