Attempt to reduce the chance of long-running DynamoDB operations (#2611)
* Add trace to DynamoDbTransactionRepository

* Use custom retry strategy for DynamoDB client

Occasionally it could take > 20 seconds to complete a DynamoDB operation
according to the x-ray traces. It's most likely due to the default retry
strategy with a maximum delay of 20 seconds.
We change to use the custom retry strategy with max delay as 1 second and
retry up to 15 times to see if it reduces the chance of long-running
DynamoDB operations.