Fix API key authorizer (#2719)
If 'x-api-key' is missing, the lambda authorizer will not be invoked and
`{ "message": "Unauthorized" }` will be returned. However, we noticed that for some
requests, `event.headers.['x-api-key']` is empty despite that it shouldn't be according
to the API Gateway logs. And we returned unauthorized policy for a valid api key
and it's cached for 50 min, before the cache expires, all the subsequent requests
using the valid api key will be unauthorized.

There could be 4 possibilites:
1. header key has upper case character (although it should be all lowercase)
2. header key has padding space
3. x-api-key is somehow put under `multiValueHeaders`
4. both `headers` and `multiValueHeaders` are somehow empty

We handle the first 3 cases and log the headers for the 4th case to see if
that's the case.