Fix double-counting issue of dashboard user stats (#3372)
* Fix double-counting issue of dashboard user stats

When refreshing the stats, we put the new values to another set
(with ready=false) first. And after it's done, we remove the old values
and remove `ready` for he new values.

Previously, as we're using `whenMatched: 'merge'` and it'll cause the following issue:

For example we only have a single user. At time t1, we refresh the dashboard stats
and get
```
{
    "_id": "2021-07",
    "drsRiskLevel_MEDIUM": 1,
}
```
because the user is MEDIUM risk at time t1.

Then at time t2, the user got changed to LOW risk. And when we refresh the dashboard again
we'll get
```
{
    "_id": "2021-07",
    "drsRiskLevel_MEDIUM": 1,
    "drsRiskLevel_LOW": 1,
}
```
resulting in double counting issue.

Other UI fixes:
- Make the user distribution charts to rotate date label
- Increase the y gap of the date label to avoid overlapping with the x-axis if 'Day' granularity
  is selected.

* fix

* fix test

* fix